TableOfContents bundled in mobile.startup
The mobile.toc module is currently conditionally loaded. That
said it's quite small. To simplify the stack and the amount of
bundles we need to manage, this component is moved into the
mobile.startup module. Follow up patches in Minerva will
remove the need for it in the codebase

Changes:
* mobile.toc.images is merged into mobile.startup.images
* template files renamed to reflect the view they are associated with
* Remove a usage of toHtmlString - icon for TableOfContents now injected
via postRender method

Bug: T207787
Change-Id: I353d6d7a6884dae03b55364150c6bb95cd4cc57f