Replace user reference with mw.user in Watchstar.js
Watchstar headless tests were failing in
Ib82b240c2582d214cce32fee5832bd28ecd21879 due to relying on a 'fresh'
import of Watchstar. Due to require() caching/test order this was not
happening:

Watchstar previously made a reference to mw.user at the top of the file.
Watchstar.test.js stubbed mw.user. However, if any of the other tests
imported Watchstar before this test ran, the Watchstar test would fail.

This commit provides a workaround to this by:

* Removing the reference to mw.user at the top of Watchstar.js and
replacing the reference with mw.user directly

In the future, we should inject user into Watchstar as a parameter to
avoid this workaround.

Change-Id: Iaf2f4de4c8564e68d20ad83c3ce1b7030ff6ec0f