Update patch set 8
Patch Set 8:

I'm not it makes sense to merge this module. I appreciate the push for fewer modules, but given that the VE code relates to a separate extension I think it makes sense to keep it separate. In practice our wikis always have VE installed, but it seems odds to load all this VE code regardless of whether it is installed or in use.

Patch-set: 8
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>