Correct the TalkBoard tests
In 50f3fdff0a TalkOverlay was rewritten as TalkBoard
but the tests were not updated. For some reason
this sneaked past Jenkins and is only making itself
known now.

Since this is blocking moves to webpack, I remove a
couple of tests relating to the talkOverlay factory
function and switch out TalkOverlay for TalkBoard

Change-Id: Ifd794779570f09753e4ce65eb7a56503f902bd51