Port mobile.categories.overlays to webpack
The categories code is lazy loaded so another entry point was added to
the webpack config and the mobile.categories.overlays module in
extension.json was modified to use this file.

Notable changes:

* another file size check was added to package.json for
mobile.categories.overlays and the mobile.common size check needed to be
increased (probably as a result of compression differences; no new
modules should have been introduced to critical js)

* mobile.categories.overlays.js was added as an entry point file and
where `m.define` statements were placed for code used by Minerva

* CategoryGateway.test.js was modified and stubbed to support headless
and browser qunit tests and to adhere to our test conventions. It's
reliance on mw.Api was also replaced with a stub.

Bug: T213472
Change-Id: I545dee3b31472ffc163c43b54375ae246a4c1347