OverlayManager: Emit overlay hide events after using #replaceCurrent
Previously, the '_om_hide' event was never emitted when an overlay
opened using #replaceCurrent was closed. Therefore #_onHideOverlay was
never called and the page URL was never updated to reflect the state.

This affected EditorOverlay when it was opened by switching from
visual editor and closed using the exit button. (It did not affect
VisualEditorOverlay, because it has no real "exit button", closing it
effectively acts the same as the browser back button.)

Bug: T209714
Bug: T213883
Change-Id: I9417c701e4b621410bca92f27f1250e07844f012