Explicitly pass in parseHTML
Make sure util runs in the context of util (don't leave it
ambiguous) and also make sure it explicitly passes in the
documentElement. This appears to be important to Safari which
is throwing a client side error, when this is not the case.

From the docs "As of 3.0 the default behavior is changed.
If the context is not specified or given as null or undefined,
a new document is used."
We don't need a new document here.

Bug: T214451
Change-Id: I81ba55278c5abaed97a8ceeb39ecae7ba659a356