MobileFrontend.skin.hooks: Cover getPluralLicenseInfo with NULL $msgObj
~ Use $this->createMock instead of $this->getMockBuilder() (hence reducing
  some lines of code).
~ Updated testGetPluralLicenseInfo() data provider method name to reflect
  what it's meant for doing.
~ Covered code path for when $msgObj is NULL. In this case, the Message
  class is not mocked but provider data given to check these cases.

NOTE: The assumption here is that "and" isn't disabled and exists in the
site language. wfMessage() creates a message object with the key "and".
Since by default, the message object uses the current user langauge as
$language = null, could there be a way to check if "and" exist in this
language without mocking the Message class?

This patch is a follow up on I771eaabd1300d4b57bbf6fc4a232e492a6e16871.

Bug: T209765
Change-Id: I844b373bbb19028e828add0dffd86d4e48306400