Update patch set 6
Patch Set 6:

(5 comments)

See inline for some minor comments.

Firstly, I think your approach with FeatureManager and Feature is great in that it neatly handles the problem of having to enable/disable features on a per-wiki basis as it's config-driven. It's certainly more flexible than anything that I've proposed in the past.

I do like the distinction between the AMC\Manager and AMC\UserMode classes. If I'm understanding your intention correctly, the former encapsulates the admin-side of enabling a mode of operation of the site and the latter encapsulates the user's ability to opt in to/out of that mode.

That being said, I think that if you were to rename UserMode to SiteMode and make the public API accept an instance of User, I think that the distinction would be unnecessary, e.g.

    AmcSiteMode::isEnabled
    AmcSiteMode::isEnabledForUser (or ::hasUserEnabled*)
    AmcSiteMode::setEnabledForUser

The latter might be considered a violation of SRP but I'd argue that it's reasonable for the case above (AMC). However, for the case of the beta mode, you might go further and create a BetaSiteModeGateway, which understands the differences in opting in/out logged in and logged out users.

This all goes to my comment on FeaturesManager::isFeatureEnabledForCurrentUser, which never actually gets the User object representing the current user…

---

* I regularly lament PHP's inability to use ? and ! in method names.

Patch-set: 6
CC: Phuedx <1196@e9e9afe9-4712-486d-8885-f54b72dd1951>