Hygiene: simplify the if statement
json_decode($x, true) can return:
 - null if $x is not proper json string
 - int, if json is a number
 - string, if json contains a string
 - array if json contains an array or object

The current code would throw warning if the passed $resp
was not an array (integer or string). The easiest, and
best way to make sure that response is correct is to
verify that is an array. There is no need to check for
null or bool

Bug: T212666
Change-Id: Ieca166a2546796d546915c87b80720615767b0b0