Include coverage reports for all files
Set the "all" flag to ensure we generate code coverage reports for
files with zero coverage... the end result a nice overview of code
health in this repo.

Begin enforcing code coverage in MobileFrontend by setting the reporting
thresholds to the current level.

Because we're enforcing code coverage this sadly means our unit tests
are run twice which is not optimal but at least ensures our tests work
in both headless and the mediawiki way.

Change-Id: Ia2fdc690a39f2280750e8988af1928512b11db34