Don't pass empty jquery array to infuseToggles
when $() searches for AMC Toggle / Beta Toggle and it doesn't find
it, it returns an empty jquery object (which is not an empty array).

This means that `if ( $('$element') ) ` will be always true. Empty
$() selector will cause OO.ui.infuse() to fail and throw exception.
Sadly, we didn't catch that because something is catching and hiding
the exception, so it looks like the Javascript executed properly.

This patch changes the if() part to make sure that the element exists
before we try to OO.ui.infuse it.

Bug: T214955
Change-Id: I139146c11f4d9211c0cdd65a3869925745df8afa