Update patch set 3
Patch Set 3: Code-Review+1

I left a suggestion inside that follows a slightly different approach. Before, the JavaScript code was compared bit by bit. But the majority of characters in this code snippet are internal implementation details and don't make any difference. The test does not need to care if, for example, spaces are used or not, or if the internally used variables are called "id" and "block".

Patch set 3 only checks interfaces that are externally used:
* The function name "mfTempOpenSection" must appear, because this is what is called in MobileFormatter::prepareHeading.
* The section ID set in MobileFormatter::createSectionBodyElement should appear.
* And the CSS class that does the toggling in the end should appear. That class is provided via several .less files.

As always this is meant as a suggestion, uploaded as a patch set on top of the existing patch for our convenience. Please feel free to update this patch however you like.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1