Fix EditorGateway test throwing an error when run in isolation
Before this commit, if the `#getContent (no section)` test from
EditorGateway.test.js was run in isolation via `QUnit.only` the
following error would appear:

```
jQuery.Deferred exception: Cannot read property 'Deferred' of undefined
```
This was happening because a promise was being executed after the test
was torn down that relied on global jQuery and our tear down step
removed the global jQuery. This commit makes the test return a promise
so that the tear down step happens after the promise is finished
instead.

Bug: T213340
Change-Id: I1083a0a1b8355bdabad1980d68b59de8dbfcfeb6