Update: add View.events support to Panels
Add new View.events API support to Panels and Drawers. This allows for
users to gracefully transition to the new API without breaking the
contract until migration is complete.

The new API allows clients to further extend View.events as subclassing
is the primary usage of Panel and Drawer today.

Bug: T208915
Change-Id: Ic193e97b4cd6f72d226bd9768e904bdec168ba39