Breaking: replace View.events with constructor args
Replace the old View.events API with a new constructor parameter,
props.events. All usages within MobileFrontend are revised in this patch
and there should be no remaining external usages.

When reviewing each module, consider:

- Verify View.events and any comments are moved to the constructor.
- Is this class extended?
  - If no, View.events can override props.events.
  - If yes, View.events must deep extend props.events since a subclass
    may wish to add events. Double check that the util class is imported
    for util.extend() usage too.
	- If changing constructor `arguments` to a named parameter such as
    props or options, verify that any Function.apply() call is replaced
    with Function.call().

This commit is prefixed "breaking" to indicate an API change. However,
there shouldn't actually be any consumers of this API so no breakage is
expected.

Bug: T208915
Change-Id: I08c0faaba50037566bef22dd401b3480081ebe60