mobile.startup/Overlay: Remove iOS 4/5 keyboard hack that no longers works
The issue this was working around is unfortunately still a problem as
of iOS 12 (opening the keyboard does not reduce the size of the
browser viewport, and instead creates an undetectable new viewport
that allow "fixed" elements to be scrolled out of view). But now there
is apparently no way to detect the size of the real viewport.

Remove some VisualEditor code that was working around the workaround.

This drops support for iOS 4 & 5 (if that even works still)
traffic for both is negligible
see https://analytics.wikimedia.org/dashboards/browsers/

Change-Id: I334e6482123578b62e5487109b95ccddc59622ef