Hygiene: allow easier access to user modes
It should be pretty easy to retrieve the single user mode to
verify is it enabled by user. The current implementation required
developer to retrieve the user mode by getting service from
MediaWikiServices (which required to know the full service name).
In the long run it would cause trouble. Developers should have
easy access to the information they often need.

Now, instead of retrieving UserMode from MediaWikiServices, or
trying to initialize the mode once again just to get the mode
identifier developer can do:

  $featureManager->getMode( 'amc' )->isEnabled();

The featureManager will be available in most places where mobile
features come into play. If feature manager is not currently
available, the second approach is to retrieve UserModes service
which will know all registered services, and it will allow to
get single service (or for example it can return all modes
enabled by user).

Below an example, how to check the user amc user mode without
having the featureManager initialized.

  MediaWikiServices::getInstance()
    ->getService( 'MobileFrontend.UserModes' )
    ->getMode( 'amc' )
    ->isEnabled();

Thanks to that, user will have to know only one service, and this
service will provide access to all registered user modes.

Additionally,
 - changed the FeautureManager class comment, as it's
not only a wrapper/collection of features, but right now it's a
facade that provides an unified access to all user modes and mobile
feautres in the MobileFrontend extension.
 - increased tests coverage for FeatureManager to 95%

Bug: T218625
Change-Id: I0a33b57b842391fbbca01c4ea3818109e0bf0418