Update patch set 4
Patch Set 4:

This is still not exactly matching, because the position of the text in read mode is also affected by the `display: table` and `vertical-align: middle` rules. But as noted in the code comment, we can't use `display: table` here, so this will have to do.

Patch-set: 4
Reviewer: Bartosz Dziewoński <417@e9e9afe9-4712-486d-8885-f54b72dd1951>