Update patch set 3
Patch Set 3:

Should there be a task associated with this/

I'm all for removing hacks but want to make sure we take the time to document why for the Wikimedia developers of the future!

Does the hack still work in iOS 4/5? If it does, we'll be breaking it there, but given browser usage, that's probably not a problem.

Patch-set: 3