Update patch set 16
Patch Set 16:

One nagging issue I have with the PromisedView constructor is that it essentially mutates the passed-in promise as a side-effect (by chaining `then`’s to it). I see the practical usage of this, because it allows us to hook onto an existing promise chain, (probably from a gateway) but I’m not sure the constructor is the place to do that. 


I thought I’d have a stab at an alternative approach, where PromisedView returns an object with a new promise and a view as properties. The downside of this though, is that we’d have to manually resolve the PromisedView promise instead of hooking onto an existing promise chain. 


https://gist.github.com/j4n-co/d9eb94570b82b87c8b919e4ee257237b 


In real life usage I’m not sure this is a great solution, but I just wanted to throw it out there.

Patch-set: 16
CC: Jdrewniak <2950@e9e9afe9-4712-486d-8885-f54b72dd1951>