Prevent editor fixed header from being scrolled out of view by touch keyboard
This workaround has already been implemented for the visual editor.
In this commit we generalize it to apply to the wikitext editor as well.

Minor change in behavior for VE: instead of using
scrollCursorIntoView() as a generic way to fix everything, we
explicitly detect and fix the one case we need to adjust. This would
be a lot harder to implement for wikitext (turns out that scrolling
the cursor into view for a <textarea> is hard, and even harder when
the textarea is unscrollable inside a scrollable wrapper, like here).

Bug: T211490
Change-Id: I579af75e451bc30dd68b9a7cc4709641f76e7213