Update patch set 3
Patch Set 3: Code-Review+2

I think this is a bug in PHPUnits coverage generation. Usually it is not a good idea to rearrange code just to work around a bug. The goal should always be to have code that is easy to read for a human being.

In this case it does not make much of a difference. The new code is fine, probably even better to read, because it avoids the long line that was split up before.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2