Do not load mobile editor on desktop, again
Allow the normal desktop editor to be used instead.

Follow-up to change 2309260e78b0d4bbac59a82e6f8b58286fa2db20, which
moved the editor init code to a module that is loaded by the Minerva
skin even on desktop. Previously the explicit check that we're in
mobile mode was not necessary.

Bug: T208068
Change-Id: If59149e569cba1b8171ddae0a39fed3fb4d0fe64