Update patch set 1
Patch Set 1:

This would be a great topic for our humans of the web meeting.

The current behaviour is intentional as it does the following
* ensures it's impossible to commit compiled JS that isn't compiled from the existing source code
* allows changes to be unstaged, which helps alleviate the need to rebase on every single commit.
* Allows you to cherry pick patches that don't change dist into other branches and then run `npm run build` and amend (sadly this seems to have backfired in your experience)

The rebase problem as been also causing problems.
One thing that's been discussed is having a bot do the building so we don't have to or making sure we commit on every patch even though that leads to rebase issues.

Marking this as WIP until we've had a conversation about the trade offs of all the options available.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>