Update patch set 17
Patch Set 17: -Code-Review

Also doesn't have to be in this patch, but what do you think about PromisedView also handling the sad path of when a promise rejects? Would it be helpful for the promise we pass to PromisedView to also reject with a View (maybe with an error message) that PromisedView would handle?

Patch-set: 17
Reviewer: Nray <6176@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0