Update patch set 5
Patch Set 5:

(14 comments)

I see several places where we use $el.find() and pass another node to it. All of them should probably simply use $(). But there is a small difference between the two ($() will just return a jQuery collection, $el.find() will also check that the parameter is a descendant of $el and return an empty collection if it's not), so they would have to be checked individually to make sure the code doesn't rely on that.

Patch-set: 5
CC: Bartosz Dziewoński <417@e9e9afe9-4712-486d-8885-f54b72dd1951>