Update patch set 4
Patch Set 4: Code-Review-1

Looks good but 1 nitpick:

Not sure if we want to tackle this here, but the spinner is not centered when loading ( and combined with the minerva patch). Looks like this is the case with master too though so it doesn't have to be a blocker. 

I think this is where it might be helpful to pull out some of the styles from languageOverlay.less and either apply the styles to both .language-overlay and talk-overlay like:

```
.language-overlay, .talk-overlay { 
  .overlay-content {
    position: relative
    min-height: 100%; 
  }
  
  .promised-view {...etc} 
} 
```
or find a way to generalize it more than that.

Patch-set: 4
Reviewer: Nray <6176@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1