Update patch set 2
Patch Set 2:

The structure test for extension.json, from MW core, is failing.

The schema for extension.json v1 is quite strict, so as to catch common mistakes (at the risk of rejecting something rare but technically valid). It currently asserts that entries in ResourceLoaderModules fit one of 5 shapes:

1. An object describing a plain file module (e.g. all known FileModule options are allowed, nothing else, including no redundant "class" key allowed).
2. An object describing a ResourceLoaderWikiModule module ('class', and the known options this constructor accepts).
3. An object describing a ResourceLoaderImageModule module ('class', and the known options this constructor accepts).
4. An object describing a custom class, defined as one with "class" not being ResourceLoaderImageModule/ResourceLoaderWikiModule/ResourceLoaderFileModule, and any free-form key-values are allowed given the class isn't known to the linter.
5. An object using a "factory" callback.

Your module here is meant to match shape number 4, but due to the class beginning with "ResourceLoaderFileModule" it is incorrectly perceived as a built-in class, and thus rejects it as plain file module shouldn't be declared in this way.

This could be worked around by renaming the class, or by improving extensions.schema.v1.json in core.

Patch-set: 2
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>