Update patch set 8
Patch Set 8:

> If you arrive on the edit page from an external link this is the only way to guarantee we leave you in a sensible place.

Interesting point, but that does not work, then. Quitting the editor after directly visiting a URL with e.g. #/editor/1 adds the fragment to the URL, but puts you at the top of the page.

Even if that worked, I think adding the fragment to the URL is confusing, and the fact that you can't use back or forward to get back into the editor is also confusing.

The task covers only restoring the section hash after save. I think we should do just that.

I agree with you that the behavior of MF overlays when visiting the page with the overlay opened from an external URL is broken and confusing. But I think we should attempt to fix that in a separate task/patch, and we should do something consistent for all overlays.

Patch-set: 8
Reviewer: Bartosz Dziewoński <417@e9e9afe9-4712-486d-8885-f54b72dd1951>