Update patch set 1
Patch Set 1: Code-Review-1

As I understand it, this would oversample all mobile edits, including the ones in the mobile wikitext editor. I actually just wanted to oversample the ones in the mobile *visual* editor, which is the place where the team is focusing all the efforts and where we don't have much data compared to the other editors.

If we do it this way, every month we will pick up an extra ~29 million mobile wikitext editor sessions which we don't really have a use for (as opposed to the 300,000 extra mobile visual editor sessions we do want).

Patch-set: 1
Reviewer: Neil P. Quinn-WMF <2518@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1