Improve overlay header building
1) Remove headerButtonsListClassName property on Overlay

The value of headerButtonsListClassName always seems to be the
same so let's make it unconfigurable and simply use header-action

2) Introduce headerActions

We need a generic way to inject "actions" that may not necessarily
be buttons into the header. On the long term this will allow us to
remove the saveHeader and previewHeader templates that are littered
across MobileFrontend. For now my use case is the addTopicOverlay
I am creating in I04bd1756a815ea70e217cf5df8a5ea6035850d86
which needs to inject a spinner into this position

Bug: T217102
Change-Id: Ie86067c6f81deebfd85efbb9537519ad7179f5a4