Generalise some textarea rules
Textareas appear inside Panel components in the talk overlay
and do not inside the EditorOverlay. These styles conflict.

This change rearranges these to avoid conflicts

The `summary` class is dropped form the talk overlay as it
appears to be unnecessary and introduces an unwanted margin
bottom inside resources/mobile.startup/Overlay.less

Bug: T217102
Change-Id: I8c34646b7ba13a26facbb69684e65109870d27a1