Remove Overlay.prototype.hasFixedHeader
The ImageOverlay is the only consumer of hasFixedHeader.
The only characteristic that makes that overlay different is that it
has no header. The only other overlay without a header is the LoadingOverlay
and that also has no need for the scrolling behaviour added in iOS.

So instead of having this property, we should derive it from the hasHeader
property.Tests are added to document how this was behaving and continue
to pass after the required updates.

The most immediate usage of this (and only thus far) is by
imageOverlay.js in I549add34101d58439b25265473a3a611c3324154.

Bug: T216198
Change-Id: I2d403a9bf5a73bab2c471a62f19e02c526e3c6fb