Let promisedView handle rejects gracefully
This commit changes promisedView so that if the promise passed in
rejects to a view (likely an error message view), promisedView will
replace its spinner view with that error view. It is opt in so that if
the promise does not reject to a view, promisedView will leave its
spinner spinning and not attempt to replace it.

The first usage of this is by ImageCarousel.js in
I549add34101d58439b25265473a3a611c3324154

Bug: T216198
Change-Id: Ia1483a908ebdd63eb0a9a2a57a566b0ab6a94922