Move creation of MobileContext to MediaWikiServices
While using MobileContext::singleton() is fine, injecting the context
is difficult when testing a method or class using the singleton approach.
So moved to service wiring so that testing this method (prepareFooter)
will be fairly simpler.

Also, cleaned up a few string quoted issues and fixed case mismatch in
a call to getLocalURL() method.

Change-Id: I806b2ac30bee2a8d965e21ad5678837409e5b230