Update patch set 1
Patch Set 1: Code-Review+2

Okay I took a look at this.

The Page object is used to store a mapping of page titles to sections that are expanded for that page.

If you visit Qatar on mobile, and expand a section it will show up here:
JSON.parse(localStorage.getItem('expandedSections')).Qatar

If you pass the table of contents this page, potentially it could create an entry to 'toc'


That said, table of contents doesn't display on mobile and localStorage only gets updated on mobile so I think it's fine to require this.

In future, we should refactor Toggler so getExpanded/saveExpanded sections are not inside these functions!

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2