Add new presentational component: carousel.js
This is a `dumb` presentational component accepting an arbitrary array
of items that a user should presumably be able to navigate through and a
featured item that is currently on display. When the user clicks one of
the navigation arrows, the onSlide callback will give you the next item
as a param! This component does nothing with the actual changing of
items - that is left up to other components.

This component will be used in I549add34101d58439b25265473a3a611c3324154
to display the image carousel. In the meantime, its .slider-button
styles will be applied to ImageOverlay (previously in Overlay.less).

Bug: T216198
Change-Id: Icce158581be7e29d6a1991815061cba634a56056