Update patch set 4
Patch Set 4: Code-Review+1

Change looks fine to me, but I've asked on the ticket that we're okay to go ahead with this.Change looks fine to me, but I've asked on the ticket that we're okay to go ahead with this.

Note you'll need to run `npm run build` to make the tests pass!

Patch-set: 4
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1