Overlays are rendered inside a container
Having them in a single container allows us to distinguish overlays
created outside the overlayManager (e.g. loadingOverlay)
from ones managed by the overlay (e.g. editor). This is helpful when
integrating with browser history as needed in T218173

The function signature for OverlayManager is changed. Given there
are no consumers in Wikimedia projects, this is not a problem.

Additional change:
* Replace a jQuery selector with a jQuery generated tag and drop
global jQuery usage

Bug: T218173
Change-Id: If4d4009a3f87b5c40ed966871659135e4fd76473