Improve carousel API
These changes were made:

* props.featured was thrown out in favor of displaying the item at
props.items[props.index].

* If props.items[prop.index] is a View, carousel will now recognize that
and render its $el property. If it is not a View, carousel will pass it
straight to JQuery's append method. Carousel should not mandate that
props.items be an array of Views because it might be desirable to have
the item you want to currently display different from the rest of the
items in the array. An example of this is by imageCarousel in
Ibd1cbc6bad2d7d3cf9fffb5e40001e107e7f3333 which will pass Carousel an
array of async functions but props.items[prop.index] will be a View
(from promisedImage).

* Tried to improving wording of jsdoc.

Bug: T216198
Change-Id: I190efb31fa5c0c8308ec880fe2ed04a61ff2af83