Update patch set 1
Patch Set 1:

> Also, if reading https://www.mediawiki.org/wiki/Skin:Minerva_Neue
right, Minerva Neue is the hard dependency for MobileFrontend,
hence do not mention Vector in the exception message.

You can run a mobile site without MobileFrontend and this is what this message is trying to make clear. You can use Timeless for example, maybe that would be a better suggestion than Vector?

See https://www.mediawiki.org/wiki/Extension:MobileFrontend#Setup_a_skin for more information. I'd love to make this message clearer!


(On subject of SkinMinerva having a hard dependency on MobileFrontend we are currently working on removing that!)

I've switched back to draft while we work this out, because clearly it's not clear enough!

Patch-set: 1
CC: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>