Do not jump to top of text editor during typing
I'm not sure what the history of the EditorOverlay scroll
onWindowScroll function is, but it seems to be in direct
conflict with the scrolling hacks disabled
in I846c0b3013074883ee39d225ed083de03fc6b0c7

After removing this code the editor appears to become usable
again. This is a great example of why the inheritance pattern is bad.
It adds lots of indirection and confusion.

Bug: T218062
Bug: T218352
Bug: T211490
Bug: T218062
Bug: T211491
Bug: T172877
Change-Id: I0781a2d1d8252c8c8004f0e8ab67705978ae3c18