Move content provider inline script to script loaded by ResourceLoader
When using the `MFContentProviderScriptPath` config option in
MobileFrontend, I have been running into `Uncaught ReferenceError: $ is
not defined` errors. Sometimes it works, but most of the time it throws
an error.

I believe this is caused by jQuery not being loaded by the time the
inline script executes. Loading this script via ResourceLoader will
ensure jQuery will be there by the time this script executes.

Change-Id: I255f11a04b015f7ae001a140e3d0dc6faf623534