18n: Reword Save to Publish buttons for mobile view sites
Per tickets such as T183901, T191104 etc, it was agreed to reword
Save to Publish on public wikis. Mobile sites are also included hence
this change.

The button to write edits to database on mobile sites is now labeled
"Publish" and the flash on success also uses the word "published" in
its text.

Bug: T191104
Change-Id: I6f6c7c139335f8b126630f31d064eb33b9393098