Improve imageDetails.js API
Intead of specifying in the jsdoc that false is required to not render
something, let's make all the children not render by default. This will
enable clients to only add keys to the props of children they actually
want to render.

Tests were modified to ensure that no children are rendered by default
which removed the need for some redundant tests.

Bug: T216198
Change-Id: I3a9d0b11bb958608775cc27a07380176b83497a1