Remove broken fade out animation on SearchOverlay
SearchOverlay extends the `hide` method of Overlay to add an animation
delay. This doesn't seem to have been working for some time as the
animations class is on the body tag.

We can restore this later using onBeforeExit if we so desire once
the OverlayManager system is more sane to deal with.

Change-Id: I01a2e46a04958b4a68650040a55f8f5ece2132aa