Refactor ImageOverlay into an Overlay with an ImageCarousel
* The onSlide method of ImageCarousel was changed to handle routing.
Previously, OverlayManager created a new instance of ImageOverlay with
each slide. Now, ImageCarousel makes a call to router and replaces
itself.

* A local reference to this.$el was created in ImageOverlay's postRender
method so that multiple promises resolving on the same slide would not
try to add an image to the same $el element. This was occurring when
rapidly advancing slides while the previous one was loading.
Additionally, in the onSlide event, a new $el reference is created and
the current instance is set to this reference to support this.

* Removes carousel.less, imageDetails.less, caorusel.js, imageDetails.js
and their associated tests that were meant for the changes related to
I549add34101d58439b25265473a3a611c3324154. Maybe these will be useful
someday, but right now my focus is to make the minimum possible changes
to remove the inheritance and satisfy T216198. The big changes
associated with I549add34101d58439b25265473a3a611c3324154 were causing
too much confusion.

* Renames ImageOverlay.hogan to ImageCarousel.hogan and changed the
extension.json to match this.

Bug: T216198
Change-Id: I3abda41ccc4a3649d7def5125c0d8c72e3505da1