Update user preferences before flusing output.
When user enables/disables the AMC or Beta mode, the database update
is performed as DeferredUpdated and performed after content is
flushed (sent to the user). This causes a race condition, as
1. user receives response which is HTTP_REDIRECT
2. server starts updating database rows for user properties
3. browser stars rendering MobileOptions page

Race condition happens between points 2 and 3, and it's possible that
browser will fetch MobileOptions page before DeferredUpdates queries
database rows. This causes an UI mismatch, as the form presents the
state before user changes, which can lead user into thinking that
options didn't change.

The simplest solution, is to still execute database queries as
DeferredUpdates, but server has to run those queries before
server returns HTTP_REDIRECT response to the user.

Bug: T219028
Change-Id: I786130a1834bd0a1bb70a75c0f6bdd312e941f5b