Only run JavaScriptTest tests locally
Running tests in Special:JavaScriptTest has proved problematic for
many reasons
1) Changes to any file update the test dist files making a merge
conflict almost certain, even if it teams are working on distinctly
different parts of the codebase
2) Developers are running the faster tests and we've been causing
avoidable problems in other extensions unit tests which have stalled
development by other teams
3) To get code coverage, we have to run the headless tests anyway.
This means we run the tests in this repo twice!
4) We're not getting any value so far out of them - not a single test
has failed in Special:JavaScript/qunit due to being run in the browser

As a result, I propose we stop committing the built assets for tests
in our commits. We add a file resources/tests.mobilefrontend.js which
is skipped and replaced with the dist file if it exists.
The skipped test allows discovery and building of tests for people who
want to run those tests in Special:JavaScript/Test. If you are a
developer you will already be running `npm start` and these tests
will be populated so I don't see this s a regression.

Change-Id: I10bbf8e625051a6de10bf4eb7fba684ae6bdf1cc