Update patch set 5
Patch Set 5:

If the problem is merge conflicts with tests (like with source code), it might be an option to not use webpack for the tests code. If I recall correctly, the only reason that was done here was due to using ../../ to access source code, which isn't necessary as demonstrated in other places. This can be done with NODE_PATH instead.

Anyhow, just a fly-by comment. Carry on.

Patch-set: 5
CC: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>