Disable browser tests on @integration with login steps
A change somewhere in the MediaWiki stack has broken all tests
in MobileFrontend with a login step. The underly method log_in doesn't
seem to be effective any more.

These tests are important so disabling them on @integration is a
temporary measure so we can merge changes to MobileFrontend
while we track down the problem

Bug: T219920
Change-Id: I15c0b15bb8ac83ea9c6a2a823f23c58a4adcd57d