Remove the Nearby component and use WatchstarPageList
Nearby is a dumb component with no need for inheritance

Nearby doesn't need to extend WatchstarPageList - it can use the
promisedView pattern to do everything it is trying to do.

postRenderLinks method which was added to ensure that the back button
after following a nearby link returned the user to their previous
scroll position is no longer working so remove the dead code.

The loading overlay usage is removed.
This cleans up usages of the LoadingOverlay by making use of
the promisedView pattern and pulling out logic hidden inside
Nearby. Two errors no longer use alerts as they can use the
MessageBox component.

The button for rendering Nearby is now disabled by default to signal
whether the button is clickable on slower connections

The gateway is updated to handle the empty list of pages case.

The resulting code is much cleaner and less mysterious than it was
before.

Bug: T217814
Change-Id: I5cb83b53e4500de169c415ae539431ad6ffd0036