Update patch set 2
Patch Set 2:

> There is no user interface for an HTTP 302 transport response.
> And you're breaking the site without keeping compat. I don't really see how there is any other option.

I'm not following. I'm not sure why there's not an option here. I can replace SpecialMobileHistory with a blank SpecialPage that redirects to the appropriate place (or leaves a link to the appropriate place). I've done that before when we migrated the editor override for action=edit. I just haven't done this yet as all I needed to do was be able to turn this page off in a development environment to showcase this proposal to my team.

Patch-set: 2
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>